add page builder and update report
