+++ .github/workflows/action_bn_ati.yml
@@ -44,12 +44,11 @@ jobs:
           python -m pip install --upgrade pip
           pip install -r scripts/requirements.txt
 
-      - name: Build BN × ATI (with DocCloud)
-        working-directory: BN_ATI_REPORT
-        run: python scripts/build_with_doccloud.py
+      - name: Build SQLite database
+        run: python scripts/build_db.py
 
       - name: Build report (writes to BN_ATI_REPORT/docs)
-        run: python scripts/build_report.py
+        run: python scripts/build_page.py
 
       - name: Ensure .nojekyll
         run: touch docs/.nojekyll
